Working - some bugs. Turns out there were some issues in factoring keyboard input routines out from output routines, passing values of stack and flag array taking too long i guess
