+++ include/output.h
@@ -1,27 +1,27 @@
+////
+//// Created by Stephen Stranahan on 2/13/2022.
+////
 //
-// Created by Stephen Stranahan on 2/13/2022.
+//#ifndef AUDIO_SYNTH_OUTPUT_H
+//#define AUDIO_SYNTH_OUTPUT_H
 //
-
-#ifndef AUDIO_SYNTH_OUTPUT_H
-#define AUDIO_SYNTH_OUTPUT_H
-
-/////////////////////////////////////////////////////////////////////////////////////////////
-////////////////////////////////////// DEFINES //////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////
-
-// Primary output signal
-#define SPKR_OUT   13
-
-/*******************************************************************************************/
-
-/////////////////////////////////////////////////////////////////////////////////////////////
-////////////////////////////////// FUNCTION PROTOS //////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////
-
-void initOutput(void);
-
-void TaskProduceOutput( void *pvParameters );
-
-/*******************************************************************************************/
-
-#endif //AUDIO_SYNTH_OUTPUT_H
+///////////////////////////////////////////////////////////////////////////////////////////////
+//////////////////////////////////////// DEFINES //////////////////////////////////////////////
+///////////////////////////////////////////////////////////////////////////////////////////////
+//
+//// Primary output signal
+//#define SPKR_OUT   13
+//
+///*******************************************************************************************/
+//
+///////////////////////////////////////////////////////////////////////////////////////////////
+//////////////////////////////////// FUNCTION PROTOS //////////////////////////////////////////
+///////////////////////////////////////////////////////////////////////////////////////////////
+//
+//void initOutput(void);
+//
+//void TaskProduceOutput( void *pvParameters );
+//
+///*******************************************************************************************/
+//
+//#endif //AUDIO_SYNTH_OUTPUT_H
